move internal/indexservice to internal/indexcoord
