+++ internal/indexcoord/util.go
